+++ src/panels/config/integrations/integration-panels/zha/zha-config-dashboard.ts
@@ -1,29 +1,25 @@
 import {
-  mdiAlertCircle,
-  mdiCheckCircle,
-  mdiCogOutline,
+  mdiAlertCircleOutline,
+  mdiCheck,
+  mdiDevices,
+  mdiDownload,
   mdiFolderMultipleOutline,
-  mdiLan,
-  mdiPencil,
+  mdiInformationOutline,
   mdiPlus,
+  mdiShape,
+  mdiTune,
 } from "@mdi/js";
 import type { CSSResultGroup, PropertyValues, TemplateResult } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
-import "../../../../../components/buttons/ha-progress-button";
 import "../../../../../components/ha-alert";
 import "../../../../../components/ha-button";
 import "../../../../../components/ha-card";
 import "../../../../../components/ha-fab";
-import "../../../../../components/ha-form/ha-form";
-import "../../../../../components/ha-icon-button";
+import "../../../../../components/ha-icon-next";
 import "../../../../../components/ha-md-list";
 import "../../../../../components/ha-md-list-item";
-import "../../../../../components/ha-select";
-import "../../../../../components/ha-textfield";
 import "../../../../../components/ha-svg-icon";
-import "../../../../../components/ha-switch";
-import type { HaSelectSelectEvent } from "../../../../../components/ha-select";
 import type { ConfigEntry } from "../../../../../data/config_entries";
 import { getConfigEntries } from "../../../../../data/config_entries";
 import type {
@@ -36,39 +32,13 @@ import {
   fetchDevices,
   fetchZHAConfiguration,
   fetchZHANetworkSettings,
-  updateZHAConfiguration,
 } from "../../../../../data/zha";
 import { showOptionsFlowDialog } from "../../../../../dialogs/config-flow/show-dialog-options-flow";
 import { showAlertDialog } from "../../../../../dialogs/generic/show-dialog-box";
-import "../../../../../layouts/hass-tabs-subpage";
-import type { PageNavigation } from "../../../../../layouts/hass-tabs-subpage";
+import { fileDownload } from "../../../../../util/file_download";
+import "../../../../../layouts/hass-subpage";
 import { haStyle } from "../../../../../resources/styles";
 import type { HomeAssistant, Route } from "../../../../../types";
-import { fileDownload } from "../../../../../util/file_download";
-import "../../../ha-config-section";
-import { showZHAChangeChannelDialog } from "./show-dialog-zha-change-channel";
-
-const MULTIPROTOCOL_ADDON_URL = "socket://core-silabs-multiprotocol:9999";
-
-const PREDEFINED_TIMEOUTS = [1800, 3600, 7200, 21600, 43200, 86400];
-
-export const zhaTabs: PageNavigation[] = [
-  {
-    translationKey: "ui.panel.config.zha.network.caption",
-    path: `/config/zha/dashboard`,
-    iconPath: mdiCogOutline,
-  },
-  {
-    translationKey: "ui.panel.config.zha.groups.caption",
-    path: `/config/zha/groups`,
-    iconPath: mdiFolderMultipleOutline,
-  },
-  {
-    translationKey: "ui.panel.config.zha.visualization.caption",
-    path: `/config/zha/visualization`,
-    iconPath: mdiLan,
-  },
-];
 
 @customElement("zha-config-dashboard")
 class ZHAConfigDashboard extends LitElement {
@@ -84,19 +54,13 @@ class ZHAConfigDashboard extends LitElement {
 
   @state() private _configuration?: ZHAConfiguration;
 
-  @state() private _networkSettings?: ZHANetworkSettings;
-
   @state() private _totalDevices = 0;
 
   @state() private _offlineDevices = 0;
 
+  @state() private _networkSettings?: ZHANetworkSettings;
+
   @state() private _error?: string;
-
-  @state() private _generatingBackup = false;
-
-  @state() private _customMains = false;
-
-  @state() private _customBattery = false;
 
   protected firstUpdated(changedProperties: PropertyValues) {
     super.firstUpdated(changedProperties);
@@ -104,81 +68,26 @@ class ZHAConfigDashboard extends LitElement {
       this.hass.loadBackendTranslation("config_panel", "zha", false);
       this._fetchConfigEntry();
       this._fetchConfiguration();
-      this._fetchSettings();
       this._fetchDevicesAndUpdateStatus();
+      this._fetchNetworkSettings();
     }
-  }
-
-  private _getUnavailableTimeoutOptions(defaultSeconds: number) {
-    const defaultSuffix = ` (${this.hass.localize("ui.panel.config.zha.configuration_page.timeout_default")})`;
-    return [
-      {
-        value: "1800",
-        label:
-          this.hass.localize(
-            "ui.panel.config.zha.configuration_page.timeout_30_min"
-          ) + (defaultSeconds === 1800 ? defaultSuffix : ""),
-      },
-      {
-        value: "3600",
-        label:
-          this.hass.localize(
-            "ui.panel.config.zha.configuration_page.timeout_1_hour"
-          ) + (defaultSeconds === 3600 ? defaultSuffix : ""),
-      },
-      {
-        value: "7200",
-        label:
-          this.hass.localize(
-            "ui.panel.config.zha.configuration_page.timeout_2_hours"
-          ) + (defaultSeconds === 7200 ? defaultSuffix : ""),
-      },
-      {
-        value: "21600",
-        label:
-          this.hass.localize(
-            "ui.panel.config.zha.configuration_page.timeout_6_hours"
-          ) + (defaultSeconds === 21600 ? defaultSuffix : ""),
-      },
-      {
-        value: "43200",
-        label:
-          this.hass.localize(
-            "ui.panel.config.zha.configuration_page.timeout_12_hours"
-          ) + (defaultSeconds === 43200 ? defaultSuffix : ""),
-      },
-      {
-        value: "86400",
-        label:
-          this.hass.localize(
-            "ui.panel.config.zha.configuration_page.timeout_24_hours"
-          ) + (defaultSeconds === 86400 ? defaultSuffix : ""),
-      },
-      {
-        value: "custom",
-        label: this.hass.localize(
-          "ui.panel.config.zha.configuration_page.timeout_custom"
-        ),
-      },
-    ];
   }
 
   protected render(): TemplateResult {
     const deviceOnline =
       this._offlineDevices < this._totalDevices || this._totalDevices === 0;
     return html`
-      <hass-tabs-subpage
+      <hass-subpage
         .hass=${this.hass}
         .narrow=${this.narrow}
-        .route=${this.route}
-        .tabs=${zhaTabs}
+        .header=${this.hass.localize("ui.panel.config.zha.network.caption")}
         back-path="/config"
         has-fab
       >
         <div class="container">
-          ${this._renderNetworkStatus(deviceOnline)} ${this._renderLightCard()}
-          ${this._renderStateCard()} ${this._renderNetworkInfoCard()}
-          ${this._renderBackupCard()} ${this._renderDynamicCards()}
+          ${this._renderNetworkStatus(deviceOnline)}
+          ${this._renderMyNetworkCard()} ${this._renderNavigationCard()}
+          ${this._renderBackupCard()}
         </div>
 
         <a href="/config/zha/add" slot="fab">
@@ -189,7 +98,7 @@ class ZHAConfigDashboard extends LitElement {
             <ha-svg-icon slot="icon" .path=${mdiPlus}></ha-svg-icon>
           </ha-fab>
         </a>
-      </hass-tabs-subpage>
+      </hass-subpage>
     `;
   }
 
@@ -201,14 +110,12 @@ class ZHAConfigDashboard extends LitElement {
           : nothing}
         <div class="card-content">
           <div class="heading">
-            <div class="icon">
+            <div class="icon ${deviceOnline ? "success" : "error"}">
               <ha-svg-icon
-                .path=${deviceOnline ? mdiCheckCircle : mdiAlertCircle}
-                class=${deviceOnline ? "online" : "offline"}
+                .path=${deviceOnline ? mdiCheck : mdiAlertCircleOutline}
               ></ha-svg-icon>
             </div>
             <div class="details">
-              Zigbee
               ${this.hass.localize(
                 `ui.panel.config.zha.configuration_page.status_${deviceOnline ? "online" : "offline"}`
               )}<br />
@@ -229,414 +136,177 @@ class ZHAConfigDashboard extends LitElement {
             </div>
           </div>
         </div>
-        <div class="card-actions">
-          <ha-button
-            href=${`/config/devices/dashboard?historyBack=1&config_entry=${this._configEntry?.entry_id}`}
-            appearance="plain"
-            size="small"
-          >
-            ${this.hass.localize("ui.panel.config.devices.caption")}</ha-button
-          >
-          <ha-button
-            appearance="plain"
-            size="small"
-            href=${`/config/entities/dashboard?historyBack=1&config_entry=${this._configEntry?.entry_id}`}
-          >
-            ${this.hass.localize("ui.panel.config.entities.caption")}</ha-button
-          >
+      </ha-card>
+    `;
+  }
+
+  private _renderMyNetworkCard() {
+    return html`
+      <ha-card class="nav-card">
+        <div class="card-header">
+          ${this.hass.localize(
+            "ui.panel.config.zha.configuration_page.my_network_title"
+          )}
+          <ha-button appearance="filled" href="/config/zha/visualization">
+            ${this.hass.localize("ui.panel.config.zha.visualization.caption")}
+          </ha-button>
+        </div>
+        <div class="card-content">
+          <ha-md-list>
+            <ha-md-list-item
+              type="link"
+              href=${`/config/devices/dashboard?historyBack=1&config_entry=${this._configEntry?.entry_id}`}
+            >
+              <ha-svg-icon slot="start" .path=${mdiDevices}></ha-svg-icon>
+              <div slot="headline">
+                ${this.hass.localize("ui.panel.config.devices.caption")}
+              </div>
+              <ha-icon-next slot="end"></ha-icon-next>
+            </ha-md-list-item>
+            <ha-md-list-item
+              type="link"
+              href=${`/config/entities/dashboard?historyBack=1&config_entry=${this._configEntry?.entry_id}`}
+            >
+              <ha-svg-icon slot="start" .path=${mdiShape}></ha-svg-icon>
+              <div slot="headline">
+                ${this.hass.localize("ui.panel.config.entities.caption")}
+              </div>
+              <ha-icon-next slot="end"></ha-icon-next>
+            </ha-md-list-item>
+            <ha-md-list-item type="link" href="/config/zha/groups">
+              <ha-svg-icon
+                slot="start"
+                .path=${mdiFolderMultipleOutline}
+              ></ha-svg-icon>
+              <div slot="headline">
+                ${this.hass.localize("ui.panel.config.zha.groups.caption")}
+              </div>
+              <ha-icon-next slot="end"></ha-icon-next>
+            </ha-md-list-item>
+          </ha-md-list>
         </div>
       </ha-card>
     `;
   }
 
-  private _renderLightCard() {
-    const data = this._configuration?.data.zha_options;
-    const transitionValue = (data?.default_light_transition as number) ?? 0;
-    return html`
-      <ha-card
-        header=${this.hass.localize(
-          "ui.panel.config.zha.configuration_page.light_card_title"
-        )}
-      >
-        ${this._configuration
-          ? html`
-              <ha-md-list>
-                <ha-md-list-item>
-                  <span slot="headline"
-                    >${this.hass.localize(
-                      "ui.panel.config.zha.configuration_page.default_light_transition_label"
-                    )}</span
-                  >
-                  <span slot="supporting-text"
-                    >${this.hass.localize(
-                      "ui.panel.config.zha.configuration_page.default_light_transition_description"
-                    )}</span
-                  >
-                  <ha-textfield
-                    slot="end"
-                    type="number"
-                    .value=${String(transitionValue)}
-                    .suffix=${"s"}
-                    .min=${0}
-                    .step=${0.5}
-                    @change=${this._defaultLightTransitionChanged}
-                  ></ha-textfield>
-                </ha-md-list-item>
-                <ha-md-list-item>
-                  <span slot="headline"
-                    >${this.hass.localize(
-                      "ui.panel.config.zha.configuration_page.enhanced_light_transition_label"
-                    )}</span
-                  >
-                  <span slot="supporting-text"
-                    >${this.hass.localize(
-                      "ui.panel.config.zha.configuration_page.enhanced_light_transition_description"
-                    )}</span
-                  >
-                  <ha-switch
-                    slot="end"
-                    .checked=${(data?.enhanced_light_transition as boolean) ??
-                    false}
-                    @change=${this._enhancedLightTransitionChanged}
-                  ></ha-switch>
-                </ha-md-list-item>
-                <ha-md-list-item>
-                  <span slot="headline"
-                    >${this.hass.localize(
-                      "ui.panel.config.zha.configuration_page.light_transitioning_flag_label"
-                    )}</span
-                  >
-                  <span slot="supporting-text"
-                    >${this.hass.localize(
-                      "ui.panel.config.zha.configuration_page.light_transitioning_flag_description"
-                    )}</span
-                  >
-                  <ha-switch
-                    slot="end"
-                    .checked=${(data?.light_transitioning_flag as boolean) ??
-                    true}
-                    @change=${this._lightTransitioningFlagChanged}
-                  ></ha-switch>
-                </ha-md-list-item>
-              </ha-md-list>
-            `
-          : nothing}
-      </ha-card>
-    `;
-  }
+  private _renderNavigationCard() {
+    const dynamicSections = this._configuration
+      ? Object.keys(this._configuration.schemas).filter(
+          (section) => section !== "zha_options"
+        )
+      : [];
 
-  private _renderStateCard() {
-    const data = this._configuration?.data.zha_options;
     return html`
-      <ha-card
-        header=${this.hass.localize(
-          "ui.panel.config.zha.configuration_page.state_card_title"
-        )}
-      >
-        ${this._configuration
-          ? html`<ha-md-list>
-              <ha-md-list-item>
-                <span slot="headline"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.enable_identify_on_join_label"
-                  )}</span
+      <ha-card class="nav-card">
+        <div class="card-content">
+          <ha-md-list>
+            <ha-md-list-item type="link" href="/config/zha/options">
+              <ha-svg-icon slot="start" .path=${mdiTune}></ha-svg-icon>
+              <div slot="headline">
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.options_title"
+                )}
+              </div>
+              <div slot="supporting-text">
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.options_description"
+                )}
+              </div>
+              <ha-icon-next slot="end"></ha-icon-next>
+            </ha-md-list-item>
+            <ha-md-list-item type="link" href="/config/zha/network-info">
+              <ha-svg-icon
+                slot="start"
+                .path=${mdiInformationOutline}
+              ></ha-svg-icon>
+              <div slot="headline">
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.network_info_title"
+                )}
+              </div>
+              <div slot="supporting-text">
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.network_info_description"
+                )}
+              </div>
+              <ha-icon-next slot="end"></ha-icon-next>
+            </ha-md-list-item>
+            ${dynamicSections.map(
+              (section) => html`
+                <ha-md-list-item
+                  type="link"
+                  href=${`/config/zha/section/${section}`}
                 >
-                <span slot="supporting-text"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.enable_identify_on_join_description"
-                  )}</span
-                >
-                <ha-switch
-                  slot="end"
-                  .checked=${(data?.enable_identify_on_join as boolean) ?? true}
-                  @change=${this._enableIdentifyOnJoinChanged}
-                ></ha-switch>
-              </ha-md-list-item>
-              <ha-md-list-item>
-                <span slot="headline"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.group_members_assume_state_label"
-                  )}</span
-                >
-                <span slot="supporting-text"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.group_members_assume_state_description"
-                  )}</span
-                >
-                <ha-switch
-                  slot="end"
-                  .checked=${(data?.group_members_assume_state as boolean) ??
-                  true}
-                  @change=${this._groupMembersAssumeStateChanged}
-                ></ha-switch>
-              </ha-md-list-item>
-              <ha-md-list-item>
-                <span slot="headline"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.consider_unavailable_mains_label"
-                  )}</span
-                >
-                <span slot="supporting-text"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.consider_unavailable_mains_description"
-                  )}</span
-                >
-                <ha-select
-                  slot="end"
-                  .value=${this._getUnavailableDropdownValue(
-                    data?.consider_unavailable_mains,
-                    this._customMains
-                  )}
-                  .options=${this._getUnavailableTimeoutOptions(7200)}
-                  @selected=${this._mainsUnavailableChanged}
-                ></ha-select>
-              </ha-md-list-item>
-              ${this._customMains
-                ? html`
-                    <ha-md-list-item>
-                      <ha-textfield
-                        slot="end"
-                        type="number"
-                        .value=${String(
-                          (data?.consider_unavailable_mains as number) ?? 7200
-                        )}
-                        .suffix=${"s"}
-                        .min=${1}
-                        .step=${1}
-                        @change=${this._customMainsSecondsChanged}
-                      ></ha-textfield>
-                    </ha-md-list-item>
-                  `
-                : nothing}
-              <ha-md-list-item>
-                <span slot="headline"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.consider_unavailable_battery_label"
-                  )}</span
-                >
-                <span slot="supporting-text"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.consider_unavailable_battery_description"
-                  )}</span
-                >
-                <ha-select
-                  slot="end"
-                  .value=${this._getUnavailableDropdownValue(
-                    data?.consider_unavailable_battery,
-                    this._customBattery
-                  )}
-                  .options=${this._getUnavailableTimeoutOptions(21600)}
-                  @selected=${this._batteryUnavailableChanged}
-                ></ha-select>
-              </ha-md-list-item>
-              ${this._customBattery
-                ? html`
-                    <ha-md-list-item>
-                      <ha-textfield
-                        slot="end"
-                        type="number"
-                        .value=${String(
-                          (data?.consider_unavailable_battery as number) ??
-                            21600
-                        )}
-                        .suffix=${"s"}
-                        .min=${1}
-                        .step=${1}
-                        @change=${this._customBatterySecondsChanged}
-                      ></ha-textfield>
-                    </ha-md-list-item>
-                  `
-                : nothing}
-              <ha-md-list-item>
-                <span slot="headline"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.enable_mains_startup_polling_label"
-                  )}</span
-                >
-                <span slot="supporting-text"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.enable_mains_startup_polling_description"
-                  )}</span
-                >
-                <ha-switch
-                  slot="end"
-                  .checked=${(data?.enable_mains_startup_polling as boolean) ??
-                  true}
-                  @change=${this._enableMainsStartupPollingChanged}
-                ></ha-switch>
-              </ha-md-list-item>
-            </ha-md-list>`
-          : nothing}
-      </ha-card>
-    `;
-  }
-
-  private _renderNetworkInfoCard() {
-    return html`
-      <ha-card
-        header=${this.hass.localize(
-          "ui.panel.config.zha.configuration_page.network_info_title"
-        )}
-      >
-        ${this._networkSettings
-          ? html`<ha-md-list class="network-info">
-              <ha-md-list-item>
-                <span slot="headline"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.channel_label"
-                  )}</span
-                >
-                <span slot="supporting-text"
-                  >${this._networkSettings.settings.network_info.channel}</span
-                >
-                <ha-icon-button
-                  slot="end"
-                  .label=${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.change_channel"
-                  )}
-                  .path=${mdiPencil}
-                  @click=${this._showChannelMigrationDialog}
-                ></ha-icon-button>
-              </ha-md-list-item>
-              <ha-md-list-item>
-                <span slot="headline">PAN ID</span>
-                <span slot="supporting-text"
-                  >${this._networkSettings.settings.network_info.pan_id}</span
-                >
-              </ha-md-list-item>
-              <ha-md-list-item>
-                <span slot="headline">Extended PAN ID</span>
-                <span slot="supporting-text"
-                  >${this._networkSettings.settings.network_info
-                    .extended_pan_id}</span
-                >
-              </ha-md-list-item>
-              <ha-md-list-item>
-                <span slot="headline">Coordinator IEEE</span>
-                <span slot="supporting-text"
-                  >${this._networkSettings.settings.node_info.ieee}</span
-                >
-              </ha-md-list-item>
-              <ha-md-list-item>
-                <span slot="headline"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.radio_type"
-                  )}</span
-                >
-                <span slot="supporting-text"
-                  >${this._networkSettings.radio_type}</span
-                >
-              </ha-md-list-item>
-              <ha-md-list-item>
-                <span slot="headline"
-                  >${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.serial_port"
-                  )}</span
-                >
-                <span slot="supporting-text"
-                  >${this._networkSettings.device.path}</span
-                >
-              </ha-md-list-item>
-              ${this._networkSettings.device.baudrate &&
-              !this._networkSettings.device.path.startsWith("socket://")
-                ? html`
-                    <ha-md-list-item>
-                      <span slot="headline"
-                        >${this.hass.localize(
-                          "ui.panel.config.zha.configuration_page.baudrate"
-                        )}</span
-                      >
-                      <span slot="supporting-text"
-                        >${this._networkSettings.device.baudrate}</span
-                      >
-                    </ha-md-list-item>
-                  `
-                : nothing}
-            </ha-md-list>`
-          : nothing}
+                  <ha-svg-icon slot="start" .path=${mdiTune}></ha-svg-icon>
+                  <div slot="headline">
+                    ${this.hass.localize(
+                      `component.zha.config_panel.${section}.title`
+                    ) || section}
+                  </div>
+                  <ha-icon-next slot="end"></ha-icon-next>
+                </ha-md-list-item>
+              `
+            )}
+          </ha-md-list>
+        </div>
       </ha-card>
     `;
   }
 
   private _renderBackupCard() {
     return html`
-      <ha-card
-        header=${this.hass.localize(
-          "ui.panel.config.zha.configuration_page.backup_restore_title"
-        )}
-      >
+      <ha-card class="nav-card">
         <div class="card-content">
-          <p>
-            ${this.hass.localize(
-              "ui.panel.config.zha.configuration_page.backup_restore_description"
-            )}
-          </p>
-        </div>
-        <div class="card-actions backup-actions">
-          <ha-progress-button
-            appearance="plain"
-            @click=${this._createAndDownloadBackup}
-            .progress=${this._generatingBackup}
-            .disabled=${!this._networkSettings || this._generatingBackup}
-          >
-            ${this.hass.localize(
-              "ui.panel.config.zha.configuration_page.download_backup"
-            )}
-          </ha-progress-button>
-          <ha-button
-            appearance="filled"
-            variant="brand"
-            @click=${this._openOptionFlow}
-          >
-            ${this.hass.localize(
-              "ui.panel.config.zha.configuration_page.migrate_radio"
-            )}
-          </ha-button>
+          <ha-md-list>
+            <ha-md-list-item>
+              <span slot="headline">
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.download_backup"
+                )}
+              </span>
+              <span slot="supporting-text">
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.download_backup_description"
+                )}
+              </span>
+              <ha-button
+                appearance="plain"
+                slot="end"
+                size="small"
+                @click=${this._createAndDownloadBackup}
+                .disabled=${!this._networkSettings}
+              >
+                <ha-svg-icon .path=${mdiDownload} slot="start"></ha-svg-icon>
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.download_backup_action"
+                )}
+              </ha-button>
+            </ha-md-list-item>
+            <ha-md-list-item>
+              <span slot="headline">
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.migrate_radio"
+                )}
+              </span>
+              <span slot="supporting-text">
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.migrate_radio_description"
+                )}
+              </span>
+              <ha-button
+                appearance="plain"
+                slot="end"
+                size="small"
+                @click=${this._openOptionFlow}
+              >
+                ${this.hass.localize(
+                  "ui.panel.config.zha.configuration_page.migrate_radio_action"
+                )}
+              </ha-button>
+            </ha-md-list-item>
+          </ha-md-list>
         </div>
       </ha-card>
-    `;
-  }
-
-  private _renderDynamicCards() {
-    if (!this._configuration) {
-      return nothing;
-    }
-    return html`
-      ${Object.entries(this._configuration.schemas)
-        .filter(([section]) => section !== "zha_options")
-        .map(
-          ([section, schema]) =>
-            html`<ha-card
-              header=${this.hass.localize(
-                `component.zha.config_panel.${section}.title`
-              )}
-            >
-              <div class="card-content">
-                <ha-form
-                  .hass=${this.hass}
-                  .schema=${schema}
-                  .data=${this._configuration!.data[section]}
-                  @value-changed=${this._dynamicDataChanged}
-                  .section=${section}
-                  .computeLabel=${this._computeLabelCallback(
-                    this.hass.localize,
-                    section
-                  )}
-                ></ha-form>
-              </div>
-              <div class="card-actions">
-                <ha-progress-button
-                  appearance="filled"
-                  variant="brand"
-                  @click=${this._updateDynamicConfiguration}
-                >
-                  ${this.hass.localize(
-                    "ui.panel.config.zha.configuration_page.update_button"
-                  )}
-                </ha-progress-button>
-              </div>
-            </ha-card>`
-        )}
     `;
   }
 
@@ -651,187 +321,14 @@ class ZHAConfigDashboard extends LitElement {
 
   private async _fetchConfiguration(): Promise<void> {
     this._configuration = await fetchZHAConfiguration(this.hass!);
-    const mainsValue = this._configuration.data.zha_options
-      ?.consider_unavailable_mains as number | undefined;
-    const batteryValue = this._configuration.data.zha_options
-      ?.consider_unavailable_battery as number | undefined;
-    this._customMains =
-      mainsValue !== undefined && !PREDEFINED_TIMEOUTS.includes(mainsValue);
-    this._customBattery =
-      batteryValue !== undefined && !PREDEFINED_TIMEOUTS.includes(batteryValue);
   }
 
-  private async _fetchSettings(): Promise<void> {
+  private async _fetchNetworkSettings(): Promise<void> {
     this._networkSettings = await fetchZHANetworkSettings(this.hass!);
   }
 
-  private async _fetchDevicesAndUpdateStatus(): Promise<void> {
-    try {
-      const devices = await fetchDevices(this.hass);
-      this._totalDevices = devices.length;
-      this._offlineDevices =
-        this._totalDevices - devices.filter((d) => d.available).length;
-    } catch (err: any) {
-      this._error = err.message || err;
-    }
-  }
-
-  private _getUnavailableDropdownValue(
-    seconds: unknown,
-    isCustom: boolean
-  ): string {
-    if (isCustom) {
-      return "custom";
-    }
-    const value = (seconds as number) ?? 7200;
-    if (PREDEFINED_TIMEOUTS.includes(value)) {
-      return String(value);
-    }
-    return "custom";
-  }
-
-  // --- Toggle handlers ---
-
-  private _groupMembersAssumeStateChanged(ev: Event): void {
-    const checked = (ev.target as HTMLInputElement).checked;
-    this._configuration!.data.zha_options.group_members_assume_state = checked;
-    this._saveConfiguration();
-  }
-
-  private _enableIdentifyOnJoinChanged(ev: Event): void {
-    const checked = (ev.target as HTMLInputElement).checked;
-    this._configuration!.data.zha_options.enable_identify_on_join = checked;
-    this._saveConfiguration();
-  }
-
-  private _enhancedLightTransitionChanged(ev: Event): void {
-    const checked = (ev.target as HTMLInputElement).checked;
-    this._configuration!.data.zha_options.enhanced_light_transition = checked;
-    this._saveConfiguration();
-  }
-
-  private _lightTransitioningFlagChanged(ev: Event): void {
-    const checked = (ev.target as HTMLInputElement).checked;
-    this._configuration!.data.zha_options.light_transitioning_flag = checked;
-    this._saveConfiguration();
-  }
-
-  private _enableMainsStartupPollingChanged(ev: Event): void {
-    const checked = (ev.target as HTMLInputElement).checked;
-    this._configuration!.data.zha_options.enable_mains_startup_polling =
-      checked;
-    this._saveConfiguration();
-  }
-
-  // --- Slider handlers ---
-
-  private _defaultLightTransitionChanged(ev: Event): void {
-    const value = Number((ev.target as HTMLInputElement).value);
-    this._configuration!.data.zha_options.default_light_transition = value;
-    this.requestUpdate();
-    this._saveConfiguration();
-  }
-
-  private _customMainsSecondsChanged(ev: Event): void {
-    const seconds = Number((ev.target as HTMLInputElement).value);
-    this._configuration!.data.zha_options.consider_unavailable_mains = seconds;
-    this.requestUpdate();
-    this._saveConfiguration();
-  }
-
-  private _customBatterySecondsChanged(ev: Event): void {
-    const seconds = Number((ev.target as HTMLInputElement).value);
-    this._configuration!.data.zha_options.consider_unavailable_battery =
-      seconds;
-    this.requestUpdate();
-    this._saveConfiguration();
-  }
-
-  // --- Select handlers ---
-
-  private _mainsUnavailableChanged(ev: HaSelectSelectEvent): void {
-    const value = ev.detail.value;
-    if (!value) {
-      return;
-    }
-    const currentValue = this._getUnavailableDropdownValue(
-      this._configuration!.data.zha_options.consider_unavailable_mains,
-      this._customMains
-    );
-    if (value === currentValue) {
-      return;
-    }
-    if (value === "custom") {
-      this._customMains = true;
-    } else {
-      this._customMains = false;
-      this._configuration!.data.zha_options.consider_unavailable_mains =
-        Number(value);
-      this._saveConfiguration();
-    }
-    this.requestUpdate();
-  }
-
-  private _batteryUnavailableChanged(ev: HaSelectSelectEvent): void {
-    const value = ev.detail.value;
-    if (!value) {
-      return;
-    }
-    const currentValue = this._getUnavailableDropdownValue(
-      this._configuration!.data.zha_options.consider_unavailable_battery,
-      this._customBattery
-    );
-    if (value === currentValue) {
-      return;
-    }
-    if (value === "custom") {
-      this._customBattery = true;
-    } else {
-      this._customBattery = false;
-      this._configuration!.data.zha_options.consider_unavailable_battery =
-        Number(value);
-      this._saveConfiguration();
-    }
-    this.requestUpdate();
-  }
-
-  // --- Save ---
-
-  private async _saveConfiguration(): Promise<void> {
-    try {
-      await updateZHAConfiguration(this.hass!, this._configuration!.data);
-    } catch (err: any) {
-      showAlertDialog(this, { text: err.message });
-    }
-  }
-
-  // --- Channel dialog ---
-
-  private async _showChannelMigrationDialog(): Promise<void> {
-    if (this._networkSettings!.device.path === MULTIPROTOCOL_ADDON_URL) {
-      showAlertDialog(this, {
-        title: this.hass.localize(
-          "ui.panel.config.zha.configuration_page.channel_dialog.title"
-        ),
-        text: this.hass.localize(
-          "ui.panel.config.zha.configuration_page.channel_dialog.text"
-        ),
-        warning: true,
-      });
-      return;
-    }
-
-    showZHAChangeChannelDialog(this, {
-      currentChannel: this._networkSettings!.settings.network_info.channel,
-    });
-  }
-
-  // --- Backup ---
-
   private async _createAndDownloadBackup(): Promise<void> {
     let backup_and_metadata: ZHANetworkBackupAndMetadata;
-
-    this._generatingBackup = true;
 
     try {
       backup_and_metadata = await createZHANetworkBackup(this.hass!);
@@ -842,8 +339,6 @@ class ZHAConfigDashboard extends LitElement {
         warning: true,
       });
       return;
-    } finally {
-      this._generatingBackup = false;
     }
 
     if (!backup_and_metadata.is_complete) {
@@ -875,33 +370,15 @@ class ZHAConfigDashboard extends LitElement {
     showOptionsFlowDialog(this, this._configEntry);
   }
 
-  // --- Dynamic card handlers (for alarm options etc.) ---
-
-  private _dynamicDataChanged(ev) {
-    this._configuration!.data[ev.currentTarget!.section] = ev.detail.value;
-  }
-
-  private async _updateDynamicConfiguration(ev: Event): Promise<void> {
-    const button = ev.currentTarget as HTMLElement & {
-      progress: boolean;
-      actionSuccess: () => void;
-      actionError: () => void;
-    };
-    button.progress = true;
+  private async _fetchDevicesAndUpdateStatus(): Promise<void> {
     try {
-      await updateZHAConfiguration(this.hass!, this._configuration!.data);
-      button.actionSuccess();
-    } catch (_err: any) {
-      button.actionError();
-    } finally {
-      button.progress = false;
+      const devices = await fetchDevices(this.hass);
+      this._totalDevices = devices.length;
+      this._offlineDevices =
+        this._totalDevices - devices.filter((d) => d.available).length;
+    } catch (err: any) {
+      this._error = err.message || err;
     }
-  }
-
-  private _computeLabelCallback(localize, section: string) {
-    return (schema) =>
-      localize(`component.zha.config_panel.${section}.${schema.name}`) ||
-      schema.name;
   }
 
   static get styles(): CSSResultGroup {
@@ -914,13 +391,19 @@ class ZHAConfigDashboard extends LitElement {
           max-width: 600px;
         }
 
-        ha-card .card-actions {
+        .nav-card .card-header {
           display: flex;
-          justify-content: flex-end;
+          align-items: center;
+          justify-content: space-between;
+          padding-bottom: var(--ha-space-2);
         }
 
-        ha-card .backup-actions {
-          justify-content: space-between;
+        .nav-card {
+          overflow: hidden;
+        }
+
+        .nav-card .card-content {
+          padding: 0;
         }
 
         .content {
@@ -936,57 +419,65 @@ class ZHAConfigDashboard extends LitElement {
           --md-item-overflow: visible;
         }
 
-        .network-info ha-md-list-item {
-          --md-list-item-supporting-text-size: var(
-            --md-list-item-label-text-size,
-            var(--md-sys-typescale-body-large-size, 1rem)
-          );
-        }
-
-        ha-select,
-        ha-textfield {
-          min-width: 210px;
-        }
-
-        @media all and (max-width: 450px) {
-          ha-select,
-          ha-textfield {
-            min-width: 160px;
-            width: 160px;
-          }
+        ha-button[size="small"] ha-svg-icon {
+          --mdc-icon-size: 16px;
         }
 
         .network-status div.heading {
           display: flex;
           align-items: center;
+          column-gap: var(--ha-space-4);
         }
 
         .network-status div.heading .icon {
-          margin-inline-end: var(--ha-space-4);
+          position: relative;
+          border-radius: var(--ha-border-radius-2xl);
+          width: 40px;
+          height: 40px;
+          display: flex;
+          align-items: center;
+          justify-content: center;
+          overflow: hidden;
+          flex-shrink: 0;
+          --icon-color: var(--primary-color);
         }
 
-        .network-status div.heading ha-svg-icon {
-          --mdc-icon-size: var(--ha-space-12);
+        .network-status div.heading .icon.success {
+          --icon-color: var(--success-color);
+        }
+
+        .network-status div.heading .icon.error {
+          --icon-color: var(--error-color);
+        }
+
+        .network-status div.heading .icon::before {
+          display: block;
+          content: "";
+          position: absolute;
+          inset: 0;
+          background-color: var(--icon-color);
+          opacity: 0.2;
+        }
+
+        .network-status div.heading .icon ha-svg-icon {
+          color: var(--icon-color);
+          width: 24px;
+          height: 24px;
         }
 
         .network-status div.heading .details {
           font-size: var(--ha-font-size-xl);
+          font-weight: var(--ha-font-weight-normal);
+          line-height: var(--ha-line-height-condensed);
+          color: var(--primary-text-color);
         }
 
         .network-status small {
           font-size: var(--ha-font-size-m);
-        }
-
-        .network-status small.offline {
+          font-weight: var(--ha-font-weight-normal);
+          line-height: var(--ha-line-height-condensed);
+          letter-spacing: 0.25px;
           color: var(--secondary-text-color);
-        }
-
-        .network-status .online {
-          color: var(--state-on-color, var(--success-color));
-        }
-
-        .network-status .offline {
-          color: var(--error-color, var(--error-color));
         }
 
         .container {
