+++ src/panels/config/integrations/integration-panels/zha/zha-groups-dashboard.ts
@@ -1,4 +1,4 @@
-import { mdiPlus } from "@mdi/js";
+import { mdiFolderMultipleOutline, mdiPlus } from "@mdi/js";
 import type { CSSResultGroup, PropertyValues, TemplateResult } from "lit";
 import { css, html, LitElement } from "lit";
 import { customElement, property, state } from "lit/decorators";
@@ -15,10 +15,18 @@ import "../../../../../components/ha-icon-button";
 import type { ZHAGroup } from "../../../../../data/zha";
 import { fetchGroups } from "../../../../../data/zha";
 import "../../../../../layouts/hass-tabs-subpage-data-table";
+import type { PageNavigation } from "../../../../../layouts/hass-tabs-subpage";
 import { haStyle } from "../../../../../resources/styles";
 import type { HomeAssistant, Route } from "../../../../../types";
 import { formatAsPaddedHex, sortZHAGroups } from "./functions";
-import { zhaTabs } from "./zha-config-dashboard";
+
+const groupsTab: PageNavigation[] = [
+  {
+    translationKey: "ui.panel.config.zha.groups.caption",
+    path: "/config/zha/groups",
+    iconPath: mdiFolderMultipleOutline,
+  },
+];
 
 export interface GroupRowData extends ZHAGroup {
   group?: GroupRowData;
@@ -100,7 +108,8 @@ export class ZHAGroupsDashboard extends LitElement {
   protected render(): TemplateResult {
     return html`
       <hass-tabs-subpage-data-table
-        .tabs=${zhaTabs}
+        .tabs=${groupsTab}
+        back-path="/config/zha/dashboard"
         .hass=${this.hass}
         .narrow=${this.narrow}
         .route=${this.route}
