+++ src/translations/en.json
@@ -6851,13 +6851,17 @@
           },
           "configuration_page": {
             "status_title": "status",
-            "status_online": "online",
-            "status_offline": "offline",
+            "status_online": "Online",
+            "status_offline": "Offline",
             "devices": "{count} {count, plural,\n one {device}\n other {devices}\n}",
             "devices_offline": "{count} offline",
             "update_button": "Update configuration",
             "download_backup": "Download backup",
+            "download_backup_description": "Save your Zigbee network configuration to a file",
+            "download_backup_action": "Download",
             "migrate_radio": "Migrate adapter",
+            "migrate_radio_description": "Move your Zigbee network to a different adapter",
+            "migrate_radio_action": "Migrate",
             "group_members_assume_state_label": "Assume state of group",
             "enable_identify_on_join_label": "Identify effect",
             "default_light_transition_label": "Light transition time",
@@ -6866,11 +6870,13 @@
             "consider_unavailable_mains_label": "Consider wired-powered devices unavailable after",
             "consider_unavailable_battery_label": "Consider battery-powered devices unavailable after",
             "enable_mains_startup_polling_label": "Refresh wired devices state on startup",
-            "network_card_title": "Network",
-            "light_card_title": "Light",
-            "state_card_title": "State",
+            "my_network_title": "My network",
+            "options_title": "Options",
+            "options_description": "Configure device behavior and network timeouts",
             "network_info_title": "Network information",
+            "network_info_description": "View channel and coordinator details",
             "backup_restore_title": "Backup and restore",
+            "backup_restore_description_short": "Download or migrate your Zigbee network backup",
             "backup_restore_description": "Back up or restore your Zigbee adapter. The backup contains your network information. It's recommended to create a backup before making any major changes to your Zigbee network.",
             "group_members_assume_state_description": "The group entity state is based on the states of its members",
             "enable_identify_on_join_description": "Show an identify effect on devices when they join the network",
@@ -6886,13 +6892,12 @@
             "serial_port": "Serial port",
             "baudrate": "Baudrate",
             "custom_seconds": "Custom",
-            "timeout_30_min": "30 minutes",
-            "timeout_1_hour": "1 hour",
-            "timeout_2_hours": "2 hours",
-            "timeout_6_hours": "6 hours",
-            "timeout_12_hours": "12 hours",
-            "timeout_24_hours": "24 hours",
-            "timeout_default": "default",
+            "timeout_30_min": "30 minutes{default}",
+            "timeout_1_hour": "1 hour{default}",
+            "timeout_2_hours": "2 hours{default}",
+            "timeout_6_hours": "6 hours{default}",
+            "timeout_12_hours": "12 hours{default}",
+            "timeout_24_hours": "24 hours{default}",
             "timeout_custom": "Custom",
             "change_channel": "Change channel",
             "channel_dialog": {
@@ -6942,7 +6947,7 @@
             "INITIALIZED_status_text": "The device is ready to use"
           },
           "network": {
-            "caption": "General"
+            "caption": "Zigbee"
           },
           "groups": {
             "add_group": "Create group",
@@ -6965,7 +6970,7 @@
             "delete": "Delete group"
           },
           "visualization": {
-            "header": "Network visualization",
+            "header": "Zigbee visualization",
             "caption": "Visualization",
             "refresh_topology": "Refresh topology",
             "device": "Device",
